🛠 Improve Interface LoggedIn History
